backend/scraper/requirements.txt
selenium
webdriver_manager
psycopg2-binary
